core: fix: invalid chunk size while creating a bytes slice
Sometimes the downloads spawn parts which are having very small size.
This would cause problem if the left amount of size is less than or
equal to the default chunk size (32KB). Hence, this fix has been implemented
to stop spawning parts further in the current part tree. Also, a condition
has been added in the parts.go copyBuffer for checking if modified chunk is
going to be greater than zero or not. This ensures more rigidity and makes
the downloader more robust.